同步分支 'origin/main' 到本地分支 'main' Updating 1234567..89abcde Fast-forward  .github/workflows/ci.yml | 2 +-  .github/workflows/release.yml | 2 +-  .signpath/artifact-configurations/windows-installer.xml | 20 ++++++++++++++++++++  LaTeXSnipper.spec | 2 +-  docs/developer_code_standards.md | 2 +-  readme.md | 2 +-  src/backend/capture_overlay.py | 2 +-  src/backend/latex_renderer.py | 2 +-  src/backend/model.py | 2 +-  src/backend/platform/macos_provider.py | 2 +-  src/backend/qhotkey/qhotkey_linux.py | 2 +-
