+++ src/backend/model.py
@@ -800,13 +800,13 @@ class ModelWrapper(QObject):
     def predict(self, pil_img: Image.Image, model_name: str = "mathcraft") -> str:
         result = self.predict_result(pil_img, model_name=model_name)
         text = str(result.get("text", "") or "").strip()
-        # 低置信度检测：如果分数极低且输出为空或疑似幻觉，给出提示
+        # Low-confidence handling: show a hint when the score is very low and output is empty or suspicious.
         score = result.get("score")
         if isinstance(score, (int, float)) and float(score) < 0.2:
             if not text:
                 return "[未识别到公式内容]"
             return f"[低置信度] {text}"
-        # 混合模式下无 score 字段，若 text 为空说明未检测到任何内容
+        # Mixed mode has no score field; empty text means nothing recognizable was detected.
         if not text:
             return "[未检测到可识别内容]"
         return text
